Patch OpenAPI spec to fix incorrect variable type

+++ pirate_weather/src/models/daily_data_inner.rs
@@ -117,7 +117,7 @@ pub struct DailyDataInner {
     pub wind_gust_time: Option<i32>,
     /// The direction of the wind in degrees.
     #[serde(rename = "windBearing", skip_serializing_if = "Option::is_none")]
-    pub wind_bearing: Option<i32>,
+    pub wind_bearing: Option<f64>,
     /// The fraction of the sky covered by clouds.
     #[serde(rename = "cloudCover", skip_serializing_if = "Option::is_none")]
     pub cloud_cover: Option<f64>,
